server/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
Flask==3.1.0
flask-cors==5.0.1
Flask-Session==0.8.0
groq==0.15.0
matplotlib==3.10.0
pymongo==4.12.0
python-dotenv==1.0.1
requests==2.25.1
spotipy==2.25.1
tqdm==4.61.2
gunicorn==23.0.0
